admin, shop api  수정

+++ src/shop/shop.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { PrismaService } from 'src/prisma.service';
 import * as bcrypt from 'bcrypt';
 
@@ -40,9 +40,59 @@ export class ShopService {
     return user;
   }
 
-  async updateUser(phoneNumber) {
-    const user = await this.prisma.uSER.delete({
-      where: {},
+  async updateUser(bodyData, res) {
+    const date = new Date();
+
+    const user = await this.prisma.uSER.findFirst({
+      where: {
+        PHONE_NUMBER: bodyData.PHONE_NUMBER,
+      },
+    });
+    if (!user) {
+      return new HttpException('NOT_JOIN', HttpStatus.FORBIDDEN);
+    }
+
+    const userUpdate = await this.prisma.uSER.update({
+      where: {
+        USER_KEY: user.USER_KEY,
+      },
+      data: {
+        NAME: bodyData.NAME,
+        PHONE_NUMBER: bodyData.PHONE_NUMBER,
+        PASSWORD: bodyData.PASSWORD,
+        UPD_ID: 'req.shop',
+        UPD_DT: date,
+      },
+    });
+
+    return res.status(200).json({
+      success: true,
+      code: 'SUCCESS',
+      msg: '업데이트 되었습니다.',
+      data: userUpdate,
+    });
+  }
+
+  async deleteUser(phoneNumber, res) {
+    const user = await this.prisma.uSER.findFirst({
+      where: {
+        PHONE_NUMBER: phoneNumber,
+      },
+    });
+    if (!user) {
+      return new HttpException('NOT_JOIN', HttpStatus.FORBIDDEN);
+    }
+
+    const deletedUser = await this.prisma.uSER.delete({
+      where: {
+        USER_KEY: user.USER_KEY,
+      },
+    });
+
+    return res.status(200).json({
+      success: true,
+      code: 'SUCCESS',
+      msg: '삭제 되었습니다.',
     });
   }
 }
